mv write_genesis back to core
